updated off-diagonal GF calculation with complex lanczos. Still not working. Need to compare with previous calculations. THE STATUS NOW IS THE FOLLOWING: -debugging lanced_tddpam_2imp VS. lanced_tddpam_bethe (1imp) at one loop the two codes give different GF.
update:
hint to debug:
1. code the true multi-band Hamiltonian case with interaction over all imp orbitals
2. solve the 1-imp case with N bath sites (N small, i.e. N=3) at 1loop and convergence
3. solve the 2-imp case with 2N bath sites with NO hybridization, decoupled Hubbard problems.
Use N bath sites per impurity by setting to zero the un-necessary link from/to imp.
This corresponds to work in the irreducible base or diagonal local hamiltonian (in this case
no rotation is required). -> the solutions of each imp must be identical and equal to the
already solve 1imp case. The off diagonal GF must be zero
4. restore the general case and turn off interaction. Self-energy must be identically zero for
all components

update2:
talked to M.Capone. Test the tpd=0 decoupled case at 1-loop setting to 0 the bath-hybridizations
mixing the impurity-bath coupling, so to mimick the decoupled bath case.