Re-factory of the code with unification of the lanc/full ed algorithm. The ED code now provides a single interface, in DMFT_ED, that can be used in f2py-->python. Main modules have been renamed using more evocative names.
fix: control variables for the Lanczos procedure have been rename.
lanc_nblock has been deleted as it is easily fixed by lanc_neigen.
The lanczos condition is now ARBITRARILY set to dim<10, for which
full diagonalization is performed instead of lanczos.

The code is ready to be extended to finite T and complex case.