update: code cleaned. H*v products have been regrouped so to have a two routines (each with  _d and _c cases): the routine spHtimesV_x perform matrix-vector product within arpack diagonalization to get the spectrum of H, using small blocks vectors in parallel mode. the routine lanc_spHtimesV_x perform matrix-vector product withing plain lanczos algorithm for green's function evaluation, using large vectors and reshuffling where necessary.
Code was tested at 1st loop against serial case and F77 code (M.Capone) with perfect
agreement.

MPI_UPDATE: mpi version of the code updated.
STORING (hash-table) of the Hamiltonian has been REINTRODUCED.
To keep memeory low the hash-table is allocated to Nchunk row on each node (previously
all nodes allocated the full sparse matrix!).
This required a modified version of the matrix mpi vector-product and to introduce a
parallel version of the matrix-vector product for the plain lanczos too (used for GF).
In addtion the building up of the Hamiltonian now (ed_geth) needs to explicitly
construct hermitian-conjugate parts of the Hamiltonian rather then obtaining them
by symmetry, in order to respect locality of the rows.

note: the products of operators needs to be normal ordered so that c-operators apply BEFORE cdg-operators.

hint: it should be possible to clean out the code by simplifying a little the definition of the matrix vector products
so to have same routines name, with same arguments!, but different behavior exploiting pre-compilation.

note2: it would be interesting to implement the nup-ndw scheme for the Hilbert space construction, while keeping
the Hamiltonian structure identical (this is a trick just for the vectors).