UPDATE: fit of the G_0(iw) added.
The flag cg_scheme has been added to select which function to fit (delta for \Delta function,
weiss for G_0). The fit scheme weiss DOES NOT rely on analytic calculation of the gradient.
While this is in princple easy to do, my implementation was kind of unstable so I decided to
rely on numerical differentiation.

Actually I extended the fmin_cg routine in SCIFOR so to behave like in scipy: it the gradient
function is optional now. If not provided is evaluated numerically.

Tests for the Hubbard model give good results, though performed for small number of Bath sites.
The Matsubara functions are almost identical, the spectral function have some differences probably
caused by the small cluster size and the different fit scheme.

Massimo noted that it is in principle more meaningful to fit the Green's function G_0, though
the \Delta function can be more sensible in some circumstances.
For the time being I will stick to \Delta fit scheme, unless the fit becomes unstable, such as
near band insulators.