update: changed the suffix in the output. Not ed_file_suffix controls a suffix placed between the filename and the actual extension of the file. In the code there are two reserved extensions: .ed for generic output, .restart for the bath output.
minor changes due to BHZ-AFM driver.

worked on the output of the code to better integrate with RDMFT.
As Giacomo noticed, calling many times in parallel the ed_solver routine
will overwrite the different outputs on the same file.
We introduced a suffix, +ed_file_suffix input variable, default is ".ed", to
each output file from the impurity solver (not the driver !!).
We took care of removing all the ".ed" from the bath input file, +Hfile input variable.
To avoid overwriting or accidental erasing.

Ready to embed into RDMFT.