subtle bug: micro bug remove in the initialization of the bath from scratch. The very small value of the hybridization vbath seemed to create problems to the cg_scheme=weiss fitting procedure, by causing the minimization to flow to large unphysical value of the parameters.
Fixed by requiring that V is at least 0.1.

Slight update of the ED_CHI2FIT.f90 module, with routines re-naming for
better understanding.
In addition re-use of the fg_delta_xxx function is implemented in the
evaluation of the fg_weiss. This ease the calculation and improve reliability.

introduced a variable ed_verbose to handle the general verbosity level of the ED code.